FIRE-8882: 2D grid snapping was bugged in the Snap to Inside case.
Turns out that if there's no axis (axis_exists == false, meaning that we are in 2D translation) then the distance from the nonexistent axis is set to 0.  0 is less than any positive number, and was forcing the 2D snapping to use the 1D snapping logic - a no-go situation.  Simply had to verify that the axis existed, and therefor that we were in 1D snapping.  Once that was done, snapping worked again.

Many thanks to Lassie for pointing out the issue.  I'd seen it during my testing but somehow had blindered myself by assuming that the error had existed before my patch…  Nope, it didn't.  My bad.

For those devs that might find the location of the opening parenthesis a little odd: I think it's odd too, however I could find nothing in the LL Coding Standard that specified how to break up larger if statements so that they are clear to read.  So I chose to follow the rules for braces: place on next line at same indent level.