FIRE-15101: Fix bottom toolbar buttons show in mouselook;
This fix reverts the initial fix for FIRE-5141 (e756e5e01e11) and replaces it
with a more specific one: The panel for the bottom toolbar buttons will always
stay visible regardless if it contains buttons or not, unless the whole toolbar
itself will be hidden (which happens for mouselook in case the UI should not be
shown). This means the nearby chat bar can be resized unless being in mouselook
and having UI in mouselook disabled.