Merge fix #2: Fix optional icon positioning in LLTextBase
Icon needs to be prepending the text if we either want it to be on the left or
the URL itself is trusted. On the other hand, only append the icon if we either
want it to be on the right side of the URL (restoring behavior prior V3.7.25)
AND if the URL itself is not trusted (we would create a duplicate trusted icon
otherwise).