FIRE-9337: Added ability to disable the FS Preprocessor on a per-script basis.
Initially I tried to use a "#pragma"-esque marker, but LSL chokes on the hash. So I moved to the simple hack of using a special-form comment.

I also started out by trying to detect the marker in the first non-whitespace line.  After some thought I realized that this might conflict with some people's editors or other things they like to add to their scripts.
To correct for this I had two options - go through all the effort of skipping commented blocks, or just simply scan the entire script for that character sequence.  I chose the latter because it is far simpler to code, even though it is more open to abuse.  The abuse, that of putting the tag in arbitrary locations in the script, can be mitigated by having the documentation state that the marker is to be located at the beginning of the script on its own line.

Because of the added scope restrictions that came about because I wrapped a block of code in an 'if' statement, I would up making the treatment of the "err" variable more consistent.  Instead of some places using the function-scoped variable and some using deeper-placed name overrides the code now just creates said variable whereever needed.