Enabled translation of the messages returned by the preprocessor code.
I made no judgement on the verbosity of the code, though it seems very noisy to me, and simply took all the strings, touched them up, and moved them out.

I initially tried to place the strings in a file more specific to the script editor, but the translation system could never find them.  So I gave up and placed them in the global string definition file.