FIRE-6786: Always show teleport location in teleport offer
This may result in the location being shown twice if it gets added to the
teleport offer automatically (default for V3 viewers, but others may not do it).
Also, if the location is being shown twice, they may differ slightly because
the actual location is being extracted from the TP lure data which is coming
from the simulator and the location in the message is calculated locally in the
other agent's viewer.

This is expected behavior until we start messing with the actual TP message which
is not really desirable as we might change what a user actually entered.