FIRE-17695: Presumed fix for being unable to drag replacement target into link replace floater on Mac/Linux
The actual cause appears to be the uninitialized variable mDeleteOnly, default to true under certain circumstances.
Also remove the unreachable code in LLFloaterLinkReplace::onTargetItemDrop() - this code will never be reached because
the method is only invoked when actually performing a drop, but in the delete case, the target item textbox
is neither visible nor enabled, so a drop can't be done at all.