FIRE-29682 - EstateChangeInfo Cap has no EEP override_environment flag
This CAP is unsued in SL (as they could not make it work (it is a TODO list item post uplift). As a result it was not updated inline with the legacy dataserver method, OpenSim does have the CAP. This fix should apply to SL when they do fix it.