[FIRE-32453] [BUG-232971] disconnect sooner to force the cache write
The call to forceQuit sets the viewer state to Quitting. This triggers all the thread pools to exit, and to do so with no sense of orderly closure. 
As a result the main window can exit which removes any visual indication to the user that work is still happening. 
When the user then restarts, issues are caused because of temporary files that may still be in use.
It is possible too that the early closure of the mainwindow, allows the OS to terminate the process without prompting the user for a safe shutdown (unproven)