Rework deployment script to unpack and support discord webhook
need to fetch the script
Make script more tolerant of errors.
Fix extraction of build_type for GitHub layout
Need to ignore pre-existing folders
We have multiple artifacts that we are merging into a common tree.
despite the clash of folders the files are not going to clash.
revert the artifact collection to per platform.
Simplify 64/32 test to remove ReleaseType dependency