Fix "duplicated copyright" error.
This solves a 'Duplicate copyrights for zlib-ng' problem.
The reason turned out to be a mysterious restart of reading the
output pipe of the command

  copyrights=autobuild('install', '--copyrights', '--install-dir', args.install_dir).

The output of that command is:

  Copyright (c) 2023, The Phoenix Firestorm Project, Inc.
  SDL2: Copyright (C) 1997-2022 Sam Lantinga
  ...
  zlib-ng: Copyright (C) 1995-2013 Jean-loup Gailly and Mark Adler

where the first line is consumed by `viewer_copyright = copyrights.readline()`.

But because the stream is read twice, it looks as if we have:

  ...
  zlib-ng: Copyright (C) 1995-2013 Jean-loup Gailly and Mark Adler
  Copyright (c) 2023, The Phoenix Firestorm Project, Inc.
  SDL2: Copyright (C) 1997-2022 Sam Lantinga
  ...
  zlib-ng: Copyright (C) 1995-2013 Jean-loup Gailly and Mark Adler

and now that first `Copyright (c) 2023, The Phoenix Firestorm Project, Inc.`
line is added to the copyright line(s) of zlib-ng (because it doesn't begin
with packagename:), which then results in "two" zlib-ng packages with
a "different" copyright line.

I am not sure if the code was buggy or that python3 on Arch is buggy,
but this code is shorter and easier to understand, too.