Allow Texture Activity to be set to Active or InActive
In order to allow a grace period while in the inactive state, the LLViewerFetchedTexture::setInactive function is modified to accept a bool that will set mTextureState to either Active or Inactive.

The setInactive function itself is then called outside of the activity and deletion check statements so the timer can be reset whenever the texture is found used on a face.