[FIRE-35024] - Store always run flag on exit and restore when logging in
Added new optional flag to restore the Always Run flag on login based on the value at exit. This is based on the existing Always Fly flag. Also added new checkbox to use the flag on login or the not(Default) in the Movement panel.

Added two new persist values to settings.xml, FSRestoreAlwaysRunAtExit and FSAlwaysRunAtExit.
Added new checkbox to panel_preferences_move.xml FSRestoreAlwaysRunAtExit, used to enable/disable the use of the AlwaysRunAtExit setting.
Updated llappviewer to save the current always run value on disconnectviewer.
Updated llagent to restore the always run flag only if the FSRestoreAlwaysRunAtExit flag is set by the user.