FIRE-35011 - Initial work on fixing the main issue (deleting textures)
Created a shadow mUUIDMap called mUUIDDeleteMap which contains a key/image pair of any object that is deleted. When a texture is deleted, it not moves over to this list, and is removed from the normal mUUIDMap and mImageList.
Currently the callbacks are still there, but not being called by not being in the main lists.
Added the reduction of calls as bias falls, adding flag to turn on/off the new feature.
Fixed issue of not disconnecting the textures on shutdown.