FIRE-35083 - Floater_stats.xml has in correct stat_view setting for materials
The issue was caused by the incorrect setting being applied to the material stat_view and child stat_bar. Added the correct named setting for the open/close state of the material stat_view and updated the float_stats.xml to have the correct settings applied. This prevents the warning and also allows the materials stat_view to have its own/close state saved correctly.