FIRE-35081 Blurry prims not changing with graphics settings, not happening with SL Viewer
Added new logic in the LLViewerTextureList::updateImageDecodePriority method to better handle the texture quality.

- Apply new rules to bias discard, there are now 2 bias, off-screen and on-screen.
        On-screen Bias
        Only applied to LOD Textures and one that have Discard > 1 (0, 1 protected)

- Off-screen Bias
        Will be using the old method of applying the mMaxVirtualSize, however
        only on LOD textures and fetched textures get bias applied.

         Local (UI & Icons), Media and Dynamic textures should not have any discard applied to them.

Without this, textures will become blurry that are on screen, which is one of the #1 user complaints.

So now, only LOD textures will change on screen, except for Fetched textures when the bias is greater then 2.00.

Also made possible fix for edge case of multiple faces for 1 texture, the last face dictates which happens to the texture. Now any that are in camera will flip signal that it is on screen.

System can be made to be be flexible.

To test out, Go to Preferences->Graphics->Rendering Settings and move the VRAM override to the far left. Then walk around and try to notice if there are any downscaled or blurry textures. To test a blurry object. Right click on it and pick edit. All of the game objects will have a max quality texture applied. Click off and see if the texture on the object changes quality to a blurred mess.