FIRE-35011 - Weird patterned extreme CPU usage when using more than 6gb vram on 10g card
Added on LLViewerFetchedTexture/LOD a raw cache which will hold the pointers for the various loaded mRawImages for the different desired discard levels.
Will clear out after 2 minutes since the last reference to not keep in memory forever.
Removed code that used old fetch workers to try to get textures that were abandoned due to transition, and instead use the raw cache.