FIRE-34977 - With 7.1.12.77180 early access update Firestorm has many freezes
Added check in LLViewerTextureList::updateImageDecodePriority to skip updating the processTextureStats if it is currently fetching, causing the thread to not stall out and cause spin locks trying to cancel and start a new fetch.
Updated LOD processTextureStats to skip checking quality of BOOST_SCULPTED textures as they are fixed to 0 and will be changed by messages from the server.
Added delay to the scaling down of the textures, where a sustained requirement of a higher discard is needed to trigger ths scale down.
Updated the LLFace check for importance to ignore the direction as when you turn the camera, want to try to reduce the amount of loading of textures.
As well forward and turning no longer set the importance to 0
Exposted delay to down scale and flag to disable texture bias in the Preferences->Graphics->Hardware Settings. Values for the delay range from 0 0ff, to 3600 seconds or 1 hour. As well as a flag to fix the bias to 1.00. Can be used by user who record video's so texture's down down scale on them.
Expose for now for testing, but could be kept.