[FIRE-35292] Fix for textures getting downscaled and compressed (J2C save fix)
Added fix for not reading enough LLImageJ2K data from the cache to save to file.
Issue is caused by hard coded data size of 999,999 used for the amount of data to read from the cache.
Used the LLImageJ2C::calcDataSizeJ2C to properly calculate the amount of data to read from the cache.