[FIRE-35184] - Atomic texture fetch states added and eased texture memory usage
Added new unordered_map for storing fetching texture states in an atomic way.
Allows for worker threads to finish without needing to be interrupted while progressing.
Allows textures to skip to the next texture that can do actual work and no longer blocks.
Added some tweaks to relax the amount of textures being kept in VRAM as well as apply the Close To Camera boost only if they are in view of the camera.

gTextureList.mFetchStates[UUID of Texture].load/store. Use those only as they can re-assign a different atomic variable to the map.