FIRE-35184 - fixed up missing std::memory_order_relaxed flag
For the Atomic values, we need to use std::memory_order_relaxed to allow them to read/write without needing to have mini spin locks as we were trying to be lock free. In the case that the updateFetch misses the value, it will take 1 loop before checking again correctly, but should be fine for performance use.
Added debug flag to support automatically using low VRAM features or on or off, used for the off screen mCloseToCamera and mInFrustum value for the textures boost.
Added default value 0 for the atomic to be reset to once the fetch is finished.