Texture VRAM Optimizations to reduce VRAM usage
LLFace.cpp
Added face to use more accurate math as change in fp:precise caused some issues with the older inaccurate math.
Tried to reduce the amount of distance for the mCloseToCamera and allow close to camera faces to be in frustrum even when off screen.

LLTextureView.cpp
Updated to show real VRAM usage (not double the amount calculated)

LLViewerDisplay.cpp
Doubled amount of time to process textures per frame

LLViewerTexture.cpp
The bytes allocated for texture and vertex were being doubled when used in the calculations so changed to more reflect how much is actually used.
Added individual texture bias and importance to camera
Testing out not changing Discard while fetching data.
Add used bias to scale the mMaxDiscardSize only when trying to get discard level.
Pulled in the max discard to not allow to go over MAX_DISCARD_LEVEL

LLViewerTextureList.cpp
Added to dump the importance to camera and bias
Fixed the bias to be 4.0 all the time
Simplified the updateImageDecodePriority to not apply bias in the calcuations
Changed to use bool values for the values instead of math using |=
Worked on some newer logic for the bias
Force LOD BOOST_NONE textures to always apply new mMaxDiscardLevel
Disabled scale down on CreateTexture from being applied to a newly loaded textures

LLVoVolume.cpp
Added small check for std::isnormal for the mPixelArea vs old_area check.