Updated LLFace to use Texture Bias (VRAM usage)
Added scaling of the distance for importance to camera and close to camera based on the current values with a an extended doubling of the distance for normal when Bias = 1.00 and goes back to the current distance when at Bias = 4.00. So it extends the distance before textures lower in quality.

Set the LLviewerTextureList::updateImageDecodePriority to use close_to_camera again, but now it shrinks in size while the bias is increasing