Texture Refactor: Smatter bias and texture scaling
This PR is designed to address issues around the scaling of textures and VRAM usage.

Added individual per texture bias. When applying bias to on screen textures, Importance to screen can shield textures from having bias applied. Scales based on the bias level.
Also modified close to camera calculation to start at 32 meters. As the bias increases, the area around the camera that is protected from the bias will decrease down to 16 meters.

Improved how textures are calculated in size so they match the actual size on screen an object is. This fixes large objects far away and small objects up close having different quality of textures applied to them.

Added some new controls for textures in the Preferences->Graphics->Hardware Settings Tab.

Auto Scale: On-Screen Textures:
On screen LOD textures will  automatically up/down scale to match the physical size of the object and no longer stick at the highest discard level. Can turn off for the previous method.

Auto-Scale: Off-Screen Textures
This allows the user to force off screen textures to have automatically scale down once off screen.

Near Off-screen Textures: Keep Texture Quality
Off-screen textures that are close to the camera will be treated as if they are on screen and close to the camera. This protects them also from bias being applied to them.

Moving Fast/Turning: Lower Texture Quality for Faster Speed
This is the existing behavior, now exposed.
While moving forward fast, or doing fast turns, all textures will have a reduction in quality to try to make loading them faster.

Removed artificial limit on VRAM calculations. VRAM that appeared in the viewer and calculations was actually 1/2 the calculated amount, due to wanting to give more wiggle room . The latest version from LL introduced a new divider to further slash the size of the VRAM, so there is no need to double dip on reducing the amount of VRAM the viewer can use. Just have the 1 new control select the divider for the VRAM...

Added new checks to only downscale a texture after 3 consecutive requests for lower quality. Any check that goes back up in quality will reset the counter.

Added new skip option for waiting for fetches until done. This is more helpful for fetches that take a long time due to slow internet download speed or issues with slow writes to disk or slow read times. It returns and will get the fetch on the next update cycle. There is a debug flag to turn this off called "FSWaitFetchDone". Didn't expose it for now.

Removed forcing all textures to be updated when ever the Bias first goes over 1.00. Depending on the # of textures loaded would cause large lag spike on the users as they were all updated in a single frame.

Reduced amount of time deleting objects per frame from 10 ms to 1 ms. Takes 10x longer, but smooths out the frames more.

Fixed scale down edge case which would try to downscale textures beyond mMaxDiscardLevel of an texture.

Added updated decode priority value for fetched textures so that the higher the boost level, the higher on the priority fetch queue.

Fixes some issues with incorrect checks being done. As well as refactored to code to remove unneeded changes and revert code back closer to original LL code to make adding changes easier. Will eventually push a PR to LL with changes.

Added additional Importance to Camera and In Frustum to textures allow for debug output in the Developer Menu Dump.

Cleaned up and updated logic for the LLViewerTextureList::updateImageDecodePriority to be up to date with latest LL version with changes reapplied.