Hide the default avatar on-rez.
Hides the default avatar until the avatar is loaded and then unmasks the default avatar selectively based upon worn mesh.

Should reduce the bleed through of linden body and the flesh-muppet cyclops.

Note: Only prevents issues due to arrival of attachments and will not fix the underlying issues with a broken COF.