Fix RelWithDebug compile issue in gltf\asset.cpp
In indra\newview\gltf\asset.cpp when you compile in Visual Studio for windows with RelWithDebug, there is a compiling error due to #if being used instead #ifdef as the rest of the codebase uses. This is with the latest Visual Studio  17.14.11.

The error given is
Severity	Code	Description	Project	File	Line	Suppression State	Details
Error (active)	E0029	expected an expression	firestorm-bin	G:\minerjr\firestorm\phoenix-firestorm\indra\newview\gltf\asset.cpp	592

Just changed #if's to #ifdef.

Also moved the second #if to #ifdef on line 611 up before the assert as in Release both checks use the attribute_mask which is not defined as its only exists when SHOW_ASSERT exists.